Port TestVariableTemplate to serial test package
